requirements.txt
requests>=2.27.1
spotipy>=2.19.0
urllib3>=1.26.8
khl.py>=0.2.11